+++ garnet-cli/src/bin/garnet.rs
@@ -259,6 +259,7 @@ fn main() -> ExitCode {
         "test" => cmd::test::run(&args[1..]),
         "fmt" => cmd::fmt::run(&args[1..]),
         "doc" => cmd::doc::run(&args[1..]),
+        "doctest" => cmd::doctest::run(&args[1..]),
         other => {
             eprintln!("unknown subcommand: {other}");
             print_help();